+++ tests/mouseMove.js
@@ -1,7 +1,7 @@
 const lepik = require("../index.js")
 
 async function main() {
-  lepik.getScreenSize().then(e => console.log(`Should print screenSize: width:${e.width} height:${e.height}`))
+  lepik.getActiveWindowId().then(console.log)
   lepik.close()
 }
 function sleep(ms) {