+++ tests/test.js
@@ -38,3 +38,4 @@ function sleep(ms) {
   return new Promise(resolve => setTimeout(resolve, ms));
 }
 main()
+